requirements.txt
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
pdfminer==20191125
pdfminer.six==20201018